# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Remote Controller Store API
  version: "0.1.0"
  description: HTTP API for the remote Juju controller store daemon (rcsd).
paths:
...
  /controllers/current:
    get:
      operationId: CurrentController
      summary: Get the current controller name
      responses:
        "200":
          description: Current controller name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NameResponse'
        "404":
          description: No current controller
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: SetCurrentController
      summary: Set the current controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameRequest'
      responses:
        "204":
          description: Updated
        "404":
          description: Controller not found
          content:
            text/plain:
              schema:
                type: string
  /controllers/previous:
    get:
      operationId: PreviousController
      summary: Get the previous controller
      responses:
        "200":
          description: Previous controller name and whether it was a switch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviousControllerInfo'
        "404":
          description: No previous controller
          content:
            text/plain:
              schema:
                type: string
  /controllers/by-endpoints:
    post:
      operationId: ControllerByAPIEndpoints
      summary: Find a controller by its API endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: Matching controller name and details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerLookupResponse'
        "404":
          description: No matching controller
          content:
            text/plain:
              schema:
                type: string
  /controllers/{name}:
    parameters:
      - $ref: '#/components/parameters/ControllerName'
    get:
      operationId: ControllerByName
      summary: Get a controller by name
      responses:
        "200":
          description: Controller details
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Not found
          content:
            text/plain:
              schema:
                type: string
    post:
      operationId: AddController
      summary: Add a new controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Created
        "409":
          description: Conflict — name or UUID already exists
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: UpdateController
      summary: Update an existing controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "204":
          description: Updated
        "404":
          description: Not found
          content:
            text/plain:
              schema:
                type: string
    delete:
      operationId: RemoveController
      summary: Remove a controller and its associated data
      responses:
        "204":
          description: Removed
...
components:
  parameters:
    ControllerName:
      name: name
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    NameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    NameResponse:
      type: object
      properties:
        name:
          type: string
    PreviousControllerInfo:
      type: object
      properties:
        name:
          type: string
        switched:
          type: boolean
    ControllerLookupResponse:
      type: object
      properties:
        name:
          type: string
        details:
          type: object
          additionalProperties: true
...